move typing context into sub module  of Ast
